restructuring of code
